+++ lib/utils/log.js
@@ -5,11 +5,11 @@ import { parentPort } from 'worker_threads';
  * @param {number} length
  */
 export function logStart(length) {
-  if (parentPort?.postMessage) {
-    parentPort.postMessage({ start: length });
-  } else {
-    // console.log(`Start: ${length}`);
-  }
+  // if (parentPort?.postMessage) {
+  //   parentPort.postMessage({ start: length });
+  // } else {
+  console.log(`Start: ${length}`);
+  // }
 }
 
 /**
