+++ lib/utils/public/data/armor-specialization.json
@@ -1,0 +1,86 @@
+[
+  {
+    "name": "Chain",
+    "slug": "chain",
+    "id": 1,
+    "type": "armor-specialization",
+    "url": "/armor-groups/1",
+    "prd": "https://2e.aonprd.com/ArmorGroups.aspx?ID=1",
+    "rarity": "common",
+    "level": null,
+    "source": [],
+    "description": "<p><strong>Source</strong> <a href=\"/sources/1\">Core Rulebook</a> pg. 275</p>\n<p>The armor is so flexible it can bend with a critical hit and absorb some of the blow. Reduce the damage from critical hits by either 4 + the value of the armor’s potency rune for medium armor, or 6 + the value of the armor’s potency rune for heavy armor. This can’t reduce the damage to less than the damage rolled for the hit before doubling for a critical hit.</p>\n"
+  },
+  {
+    "name": "Composite",
+    "slug": "composite",
+    "id": 2,
+    "type": "armor-specialization",
+    "url": "/armor-groups/2",
+    "prd": "https://2e.aonprd.com/ArmorGroups.aspx?ID=2",
+    "rarity": "common",
+    "level": null,
+    "source": [],
+    "description": "<p><strong>Source</strong> <a href=\"/sources/1\">Core Rulebook</a> pg. 275</p>\n<p>The numerous overlapping pieces of this armor protect you from piercing attacks. You gain resistance to piercing damage equal to 1 + the value of the armor’s potency rune for medium armor, or 2 + the value of the armor’s potency rune for heavy armor.</p>\n"
+  },
+  {
+    "name": "Leather",
+    "slug": "leather",
+    "id": 3,
+    "type": "armor-specialization",
+    "url": "/armor-groups/3",
+    "prd": "https://2e.aonprd.com/ArmorGroups.aspx?ID=3",
+    "rarity": "common",
+    "level": null,
+    "source": [],
+    "description": "<p><strong>Source</strong> <a href=\"/sources/1\">Core Rulebook</a> pg. 275</p>\n<p>The thick second skin of the armor disperses blunt force to reduce bludgeoning damage. You gain resistance to bludgeoning damage equal to 1 + the value of the armor’s potency rune for medium armor, or 2 + the value of the armor’s potency rune for heavy armor.</p>\n"
+  },
+  {
+    "name": "Plate",
+    "slug": "plate",
+    "id": 4,
+    "type": "armor-specialization",
+    "url": "/armor-groups/4",
+    "prd": "https://2e.aonprd.com/ArmorGroups.aspx?ID=4",
+    "rarity": "common",
+    "level": null,
+    "source": [],
+    "description": "<p><strong>Source</strong> <a href=\"/sources/1\">Core Rulebook</a> pg. 275</p>\n<p>The sturdy plate provides no purchase for a cutting edge. You gain resistance to slashing damage equal to 1 + the value of the armor’s potency rune for medium armor, or 2 + the value of the armor’s potency rune for heavy armor.</p>\n"
+  },
+  {
+    "name": "Cloth",
+    "slug": "cloth",
+    "id": 5,
+    "type": "armor-specialization",
+    "url": "/armor-groups/5",
+    "prd": "https://2e.aonprd.com/ArmorGroups.aspx?ID=5",
+    "rarity": "common",
+    "level": null,
+    "source": [],
+    "description": "<p><strong>Source</strong> <a href=\"/sources/1\">Core Rulebook</a> pg. 275</p>\n<p><em>Nethys Note: there is no medium or heavy cloth armor, so there is no armor specialization effect for the cloth armor group.</em></p>\n"
+  },
+  {
+    "name": "Skeletal",
+    "slug": "skeletal",
+    "id": 6,
+    "type": "armor-specialization",
+    "url": "/armor-groups/6",
+    "prd": "https://2e.aonprd.com/ArmorGroups.aspx?ID=6",
+    "rarity": "common",
+    "level": null,
+    "source": [],
+    "description": "<p><strong>Source</strong> <a href=\"/sources/191\">Treasure Vault</a> pg. 8</p>\n<p>Armor made from the bone or exoskeleton of creatures as diverse as bears, insects, and coral, skeletal armor protects vital points from precision damage. You gain resistance to precision damage equal to 3 + the value of the armor's potency rune for medium armor, or 5 + the value of the armor's potency rune for heavy armor</p>\n"
+  },
+  {
+    "name": "Wood",
+    "slug": "wood",
+    "id": 7,
+    "type": "armor-specialization",
+    "url": "/armor-groups/7",
+    "prd": "https://2e.aonprd.com/ArmorGroups.aspx?ID=7",
+    "rarity": "common",
+    "level": null,
+    "source": [],
+    "description": "<p><strong>Source</strong> <a href=\"/sources/191\">Treasure Vault</a> pg. 9</p>\n<p>Wood armor is generally flexible and light, but it can splinter as it breaks, throwing off shards and fragments that damage foes who deal you critical blows. If a foe critically hits you with a melee unarmed attack or critically hits you with any melee attack while adjacent to you, it takes piercing damage equal to 3 + the armor's potency rune value for medium armor, or 5 + the armor's potency rune value for heavy armor.</p>\n"
+  }
+]
